# contracts/openapi/openapi-v1.yaml
openapi: 3.1.0
info:
  title: CppDefense API
  version: 1.1.0
  summary: Public application and private runner contracts for CppDefense 2.0
  license:
    name: Proprietary source-available
servers:
  - url: https://cppdefense.example
    description: Placeholder production origin
...
paths:
...
  /api/v1/auth/github/start:
    get:
      tags: [Auth]
      operationId: startGithubLogin
      parameters:
        - $ref: '#/components/parameters/ReturnTo'
      responses:
        '302':
          description: Redirect to GitHub OAuth authorization endpoint
          headers:
            Location:
              schema:
                type: string
                format: uri
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/auth/github/callback:
    get:
      tags: [Auth]
      operationId: completeGithubLogin
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - name: state
          in: query
          required: true
          schema:
            type: string
            minLength: 32
            maxLength: 512
      responses:
        '302':
          description: Login completed; redirect to application or pending screen
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/BadGateway'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      security:
        - cookieAuth: []
          csrfToken: []
      responses:
        '204':
          description: Session revoked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/me:
    get:
      tags: [Profile]
      operationId: getCurrentUser
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/groups:
    get:
      tags: [Groups]
      operationId: listGroups
      x-roles: [teacher, admin]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Groups visible to current user
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
                  next_cursor:
                    type: [string, 'null']
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Groups]
      operationId: createGroup
      x-roles: [admin]
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code, name]
              properties:
                code:
                  $ref: '#/components/schemas/Code'
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                is_demo:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/groups/{group_id}/labs:
    get:
      tags: [Groups]
      operationId: listGroupLabs
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Labs visible to current user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lab'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Groups]
      operationId: createLab
      x-roles: [teacher, admin]
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLabRequest'
      responses:
        '201':
          description: Lab created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lab'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/groups/{group_id}/submissions:
    get:
      tags: [Groups]
      operationId: listGroupSubmissions
      x-roles: [teacher, admin]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: Immutable submission versions in a managed group
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/GroupSubmission'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/submissions/{submission_id}/archive:
    get:
      tags: [Groups]
      operationId: downloadSubmissionArchive
      security:
        - cookieAuth: []
      parameters:
        - name: submission_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: ZIP visible to its student, assigned teacher, or administrator
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/imports:
    post:
      tags: [Imports]
      operationId: createImport
      x-roles: [teacher, admin]
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [group_id, kind, archive]
              properties:
                group_id:
                  $ref: '#/components/schemas/Uuid'
                kind:
                  enum: [group, lab]
                archive:
                  type: string
                  format: binary
      responses:
        '202':
          description: Upload accepted for validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Import'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/imports/{import_id}:
    get:
      tags: [Imports]
      operationId: getImport
      x-roles: [teacher, admin]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ImportId'
      responses:
        '200':
          description: Import and preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Import'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/imports/{import_id}/review:
    post:
      tags: [Imports]
      operationId: reviewImport
      x-roles: [teacher, admin]
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/ImportId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [decision, archive_sha256, checklist]
              properties:
                decision:
                  enum: [approve, reject]
                archive_sha256:
                  $ref: '#/components/schemas/Sha256'
                checklist:
                  $ref: '#/components/schemas/ReviewChecklist'
                reason:
                  type: string
                  maxLength: 2000
      responses:
        '200':
          description: Review stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Import'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/imports/{import_id}/apply:
    post:
      tags: [Imports]
      operationId: applyImport
      x-roles: [teacher, admin]
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/ImportId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Approved import application started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Import'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/defenses:
    post:
      tags: [Defenses]
      operationId: createDefense
      x-roles: [student]
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [submission_version_id]
              properties:
                submission_version_id:
                  $ref: '#/components/schemas/Uuid'
      responses:
        '202':
          description: Defense created and preparation queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Defense'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/defenses/{defense_id}:
    get:
      tags: [Defenses]
      operationId: getDefense
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/DefenseId'
      responses:
        '200':
          description: Defense visible to current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Defense'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/defenses/{defense_id}/draft:
    put:
      tags: [Defenses]
      operationId: saveDefenseDraft
      x-roles: [student]
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/DefenseId'
        - name: If-Match
          in: header
          required: true
          schema:
            type: string
            pattern: '^"[0-9]+"$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [answer]
              properties:
                answer:
                  type: string
                  maxLength: 1048576
      responses:
        '200':
          description: Draft saved
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [version, saved_at]
                properties:
                  version:
                    type: integer
                    minimum: 1
                  saved_at:
                    type: string
                    format: date-time
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/defenses/{defense_id}/attempts:
    post:
      tags: [Defenses]
      operationId: createCheckAttempt
      x-roles: [student]
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/DefenseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [answer]
              properties:
                answer:
                  type: string
                  maxLength: 1048576
      responses:
        '202':
          description: Attempt and job created atomically
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckAttempt'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/defenses/{defense_id}/cancel:
    post:
      tags: [Defenses]
      operationId: cancelDefense
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/DefenseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Defense and outstanding jobs cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Defense'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/attempts/{attempt_id}:
    get:
      tags: [Defenses]
      operationId: getCheckAttempt
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/AttemptId'
      responses:
        '200':
          description: Attempt visible under object policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckAttempt'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/history/attempts:
    get:
      tags: [History]
      operationId: listAttemptHistory
      x-roles: [teacher, admin]
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - name: group_id
          in: query
          schema:
            $ref: '#/components/schemas/Uuid'
        - name: student_record_id
          in: query
          schema:
            $ref: '#/components/schemas/Uuid'
        - name: lab_id
          in: query
          schema:
            $ref: '#/components/schemas/Uuid'
        - name: outcome
          in: query
          schema:
            $ref: '#/components/schemas/AttemptOutcome'
      responses:
        '200':
          description: Scoped attempt history
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CheckAttempt'
                  next_cursor:
                    type: [string, 'null']
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/pending-links:
    get:
      tags: [Administration]
      operationId: listPendingStudentLinks
      x-roles: [teacher, admin]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Pending GitHub accounts with scoped matching student records
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      additionalProperties: false
                      required: [user_id, github_login]
                      properties:
                        user_id: { $ref: '#/components/schemas/Uuid' }
                        display_name: { type: [string, 'null'] }
                        github_login: { type: string }
                        student_record_id:
                          oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }]
                        group_id:
                          oneOf: [{ $ref: '#/components/schemas/Uuid' }, { type: 'null' }]
                        group_code: { type: [string, 'null'] }
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/pending-links/{user_id}/approve:
    post:
      tags: [Administration]
      operationId: approvePendingStudentLink
      x-roles: [teacher, admin]
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [student_record_id]
              properties:
                student_record_id:
                  $ref: '#/components/schemas/Uuid'
                reason:
                  type: string
                  maxLength: 2000
      responses:
        '200':
          description: User activated and linked atomically
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/pending-links/{user_id}/reject:
    post:
      tags: [Administration]
      operationId: rejectPendingStudentLink
      x-roles: [teacher, admin]
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 1
                  maxLength: 2000
      responses:
        '200':
          description: Pending association rejected and audited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/users:
    get:
      tags: [Administration]
      operationId: listManagedUsers
      x-roles: [admin]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Users visible to the administrator
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ManagedUser'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/users/{user_id}/role:
    put:
      tags: [Administration]
      operationId: setUserRole
      x-roles: [admin]
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role]
              properties:
                role:
                  enum: [student, teacher, admin]
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedUser'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/users/{user_id}/block:
    post:
      tags: [Administration]
      operationId: blockUser
      x-roles: [admin]
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 1
                  maxLength: 2000
      responses:
        '200':
          description: User blocked and active sessions revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagedUser'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/groups/{group_id}/teachers:
    post:
      tags: [Administration]
      operationId: assignTeacherToGroup
      x-roles: [admin]
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [user_id]
              properties:
                user_id:
                  $ref: '#/components/schemas/Uuid'
      responses:
        '204':
          description: Teacher assigned
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/view-as-role:
    put:
      tags: [Administration]
      operationId: setAdminViewAsRole
      x-roles: [admin]
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ViewAsRoleRequest'
      responses:
        '200':
          description: Session presentation mode changed; real role remains admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [Administration]
      operationId: clearAdminViewAsRole
      x-roles: [admin]
      security:
        - cookieAuth: []
          csrfToken: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Session presentation mode reset to admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/runner/jobs/lease:
    post:
      tags: [Runner]
      operationId: leaseJob
      security:
        - runnerBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [runner_id, capacity]
              properties:
                runner_id:
                  $ref: '#/components/schemas/Uuid'
                capacity:
                  type: integer
                  minimum: 1
                  maximum: 6
      responses:
        '200':
          description: One job leased
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobLease'
        '204':
          description: No eligible job
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/runner/jobs/{job_id}/heartbeat:
    post:
      tags: [Runner]
      operationId: heartbeatJob
      security:
        - runnerBearer: []
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/LeaseToken'
      responses:
        '204':
          description: Lease extended
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/runner/jobs/{job_id}/complete:
    post:
      tags: [Runner]
      operationId: completeJob
      security:
        - runnerBearer: []
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/LeaseToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteJobRequest'
      responses:
        '200':
          description: Current lease result accepted idempotently
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckAttempt'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cppdefense_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
    runnerBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-service-token
  parameters:
    ReturnTo:
      name: return_to
      in: query
      required: false
      schema:
        type: string
        maxLength: 2048
        pattern: '^/(?!/).*$'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Authenticated principal is not allowed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Conflict:
      description: State, version or idempotency conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    BadGateway:
      description: Upstream identity provider failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Uuid:
      type: string
      format: uuid
...
    Code:
      type: string
      minLength: 1
      maxLength: 64
      pattern: '^[a-z0-9](?:[a-z0-9-]{0,62}[a-z0-9])?$'
...